上周解决的问题
产品查询窗口，增加确认按钮
scheduled 2012年12月22日 from 下午2:30 to 下午2:45

OA特价单记录类型错误
scheduled 2012年12月21日 from 下午12:45 to 下午1:00

审批actions标签汉化
scheduled 2012年12月21日 from 上午11:30 to 上午11:45

交货周期同步
scheduled 2012年12月21日 from 下午12:45 to 下午1:00

单身行数必须大于零
scheduled 2012年12月21日 from 下午1:30 to 下午2:00

字段长度确认
scheduled 2012年12月22日 from 下午3:30 to 下午3:45

审批Approval标签汉化
scheduled 2012年12月21日 from 上午11:45 to 下午12:00

订单删除问题
scheduled 2012年12月22日 from 下午9:00 to 下午9:15

sharerole 字段限制
scheduled 2012年12月21日 from 下午9:30 to 下午9:45

取消回车新增行
scheduled 2012年12月21日 from 上午10:45 to 上午11:15

数量输入问题
scheduled 2012年12月22日 from 下午6:15 to 下午6:30

菜单加载问题
scheduled 2012年12月22日 from 上午10:45 to 上午11:30

OA特价单折扣确认
scheduled 2012年12月22日 from 下午4:30 to 下午4:45

订单单头增加最终金额合计
scheduled 2012年12月21日 from 下午1:15 to 下午1:30

报表分页
scheduled 2012年12月22日 from 下午1:15 to 下午1:30

如果产品没有属性，则默认允许零价格
scheduled 2012年12月21日 from 下午2:00 to 下午2:15

单身为零提示
scheduled 2012年12月21日 from 下午12:15 to 下午12:30

订单单身删除提示
scheduled 2012年12月21日 from 下午1:00 to 下午1:15

单身数量必须大于零
scheduled 2012年12月21日 from 上午10:30 to 上午10:45

小于3K
scheduled 2012年12月22日 from 下午12:00 to 下午12:00

列表默认30行
scheduled 2012年12月21日 from 下午9:00 to 下午9:15

订单删除行后，主附件全部展开
scheduled 2012年12月22日 from 下午7:15 to 下午7:30

项目esc问题
scheduled 2012年12月21日 from 下午10:00 to 下午10:15

显示订单时，提示是否要提交
scheduled 2012年12月22日 from 下午8:00 to 下午8:15

回车添加单身问题
scheduled 2012年12月21日 from 下午12:15 to 下午12:45

+++ eos2/grails-app/controllers/org/leaf/eos2/b2b/SalesOrder224Controller.groovy
@@ -3,6 +3,7 @@ package org.leaf.eos2.b2b
 class SalesOrder224Controller {
 
 	def salesOrder224Service
+	def salesOrderService
 
     def index = { redirect(action: "list", params: params) }
 
@@ -10,8 +11,8 @@ class SalesOrder224Controller {
     static allowedMethods = [save: "POST", update: "POST", delete: "POST"]
 
     def list = {
-        params.max = Math.min(params.max ? params.max.toInteger() : 10,  100)
-        render view:'/b2b/salesOrder/list', model:[salesOrderInstanceList: SalesOrder.list(params), salesOrderInstanceTotal: SalesOrder.count()]
+        params.max = Math.min(params.max ? params.max.toInteger() : 30,  100)
+        render view:'/b2b/salesOrder/list', model:[salesOrderInstanceList: salesOrderService.list(params), salesOrderInstanceTotal: salesOrderService.count(params)]
     }
 
     def create = {
@@ -33,7 +34,7 @@ class SalesOrder224Controller {
     }
 
     def show = {
-        def salesOrderInstance = SalesOrder.get(params.id)
+        def salesOrderInstance = salesOrderService.show(params.id)
         if (!salesOrderInstance) {
             flash.message = "salesOrder.not.found"
             flash.args = [params.id]
@@ -46,7 +47,12 @@ class SalesOrder224Controller {
     }
 
     def edit = {
-        def salesOrderInstance = SalesOrder.get(params.id)
+		if(salesOrderService.isLocked(params.id) == true){
+			render(view : "/unauthorized") 
+            return
+		}
+
+        def salesOrderInstance = salesOrderService.edit(params.id)
         if (!salesOrderInstance) {
             flash.message = "salesOrder.not.found"
             flash.args = [params.id]
@@ -59,7 +65,12 @@ class SalesOrder224Controller {
     }
 
     def update = {
-        def salesOrderInstance = SalesOrder.get(params.id)
+		if(salesOrderService.isLocked(params.id) == true){
+			render(view : "/unauthorized") 
+            return
+		}
+
+        def salesOrderInstance = salesOrderService.edit(params.id)
         if (salesOrderInstance) {
             if (params.version) {
                 def version = params.version.toLong()
@@ -90,7 +101,12 @@ class SalesOrder224Controller {
     }
 
     def delete = {
-        def salesOrderInstance = SalesOrder.get(params.id)
+		if(salesOrderService.isLocked(params.id) == true){
+			render(view : "/unauthorized") 
+            return
+		}
+
+        def salesOrderInstance = salesOrderService.delete(params.id)
         if (salesOrderInstance) {
             try {
                 salesOrderInstance.delete()
